move use of generator store out of ExportScriptDialog
