+++ src/views/Generator/Generator.hooks.ts
@@ -2,12 +2,13 @@ import { useParams } from 'react-router-dom'
 import invariant from 'tiny-invariant'
 
 import { useToast } from '@/store/ui/useToast'
-import { loadGeneratorFile, loadHarFile } from './Generator.utils'
+import { exportScript, loadGeneratorFile, loadHarFile } from './Generator.utils'
 import { selectGeneratorData, useGeneratorStore } from '@/store/generator'
 import { GeneratorFileData } from '@/types/generator'
 import { useMutation, useQuery } from '@tanstack/react-query'
 import { queryClient } from '@/utils/query'
 import log from 'electron-log/renderer'
+import { useCallback } from 'react'
 
 export function useGeneratorParams() {
   const { fileName } = useParams()
@@ -87,5 +88,41 @@ export function useIsGeneratorDirty(fileName: string) {
   // `property: undefined` values
   return (
     JSON.stringify(generatorStateData) !== JSON.stringify(generatorFileData)
+  )
+}
+
+export function useScriptExport(generatorFileName: string) {
+  const showToast = useToast()
+  const setScriptName = useGeneratorStore((store) => store.setScriptName)
+  const { mutateAsync: updateGeneratorFile } =
+    useUpdateValueInGeneratorFile(generatorFileName)
+
+  return useCallback(
+    async (scriptName: string) => {
+      try {
+        await exportScript(scriptName)
+      } catch (error) {
+        log.error(error)
+
+        showToast({
+          title: 'Failed to export script',
+          status: 'error',
+        })
+      }
+
+      try {
+        await updateGeneratorFile({ key: 'scriptName', value: scriptName })
+
+        setScriptName(scriptName)
+      } catch (error) {
+        log.error(error)
+
+        showToast({
+          title: 'Failed to update script name',
+          status: 'error',
+        })
+      }
+    },
+    [showToast, setScriptName, updateGeneratorFile]
   )
 }
