+++ src/views/Generator/GeneratorControls/GeneratorControls.tsx
@@ -2,21 +2,16 @@ import { useState } from 'react'
 import { DropdownMenu, Flex, IconButton } from '@radix-ui/themes'
 
 import { useScriptPreview } from '@/hooks/useScriptPreview'
-import { exportScript } from '../Generator.utils'
 import { ValidatorDialog } from '../ValidatorDialog'
 import { ExportScriptDialog } from '../ExportScriptDialog'
 import { DotsVerticalIcon } from '@radix-ui/react-icons'
-import {
-  useGeneratorParams,
-  useUpdateValueInGeneratorFile,
-} from '../Generator.hooks'
+import { useGeneratorParams, useScriptExport } from '../Generator.hooks'
 import { useNavigate } from 'react-router-dom'
 import { getRoutePath } from '@/routeMap'
 import { ButtonWithTooltip } from '@/components/ButtonWithTooltip'
 import { getFileNameWithoutExtension } from '@/utils/file'
 import { RecordingSelector } from '../RecordingSelector'
-import { useToast } from '@/store/ui/useToast'
-import log from 'electron-log/renderer'
+import { useGeneratorStore } from '@/store/generator'
 
 interface GeneratorControlsProps {
   onSave: () => void
@@ -24,6 +19,8 @@ interface GeneratorControlsProps {
 }
 
 export function GeneratorControls({ onSave, isDirty }: GeneratorControlsProps) {
+  const scriptName = useGeneratorStore((store) => store.scriptName)
+
   const [isValidatorDialogOpen, setIsValidatorDialogOpen] = useState(false)
   const [isExportScriptDialogOpen, setIsExportScriptDialogOpen] =
     useState(false)
@@ -31,11 +28,6 @@ export function GeneratorControls({ onSave, isDirty }: GeneratorControlsProps) {
   const { preview, hasError } = useScriptPreview()
   const isScriptExportable = !hasError && !!preview
   const navigate = useNavigate()
-
-  const { mutateAsync: updateGeneratorFile } =
-    useUpdateValueInGeneratorFile(fileName)
-
-  const showToast = useToast()
 
   const handleDeleteGenerator = async () => {
     await window.studio.ui.deleteFile({
@@ -47,29 +39,7 @@ export function GeneratorControls({ onSave, isDirty }: GeneratorControlsProps) {
     navigate(getRoutePath('home'))
   }
 
-  const handleExportScript = async (fileName: string) => {
-    try {
-      await exportScript(fileName)
-    } catch (error) {
-      log.error(error)
-
-      showToast({
-        title: 'Failed to export script',
-        status: 'error',
-      })
-    }
-
-    try {
-      await updateGeneratorFile({ key: 'scriptName', value: fileName })
-    } catch (error) {
-      log.error(error)
-
-      showToast({
-        title: 'Failed to update script name',
-        status: 'error',
-      })
-    }
-  }
+  const handleExportScript = useScriptExport(fileName)
 
   return (
     <>
@@ -116,6 +86,7 @@ export function GeneratorControls({ onSave, isDirty }: GeneratorControlsProps) {
             />
             <ExportScriptDialog
               open={isExportScriptDialogOpen}
+              scriptName={scriptName}
               onExport={handleExportScript}
               onOpenChange={setIsExportScriptDialogOpen}
             />
